--- a/assets/js/aws.permissions.cloud.js
+++ b/assets/js/aws.permissions.cloud.js
@@ -131,20 +131,20 @@ function getTemplates(action, iam_def) {
     return ret;
 }
 
-let privilege_2_iam_mapping_name = null;
+let privilege_to_iam_mapping_name = null;
 function getUsedBy(privilege, sdk_map) {
-    if (privilege_2_iam_mapping_name === null) {
-        privilege_2_iam_mapping_name = {}
+    if (privilege_to_iam_mapping_name === null) {
+        privilege_to_iam_mapping_name = {}
         for (let iam_mapping_name of Object.keys(sdk_map['sdk_method_iam_mappings']).sort()) {
             for (let action of sdk_map['sdk_method_iam_mappings'][iam_mapping_name]) {
-                if (privilege_2_iam_mapping_name[action['action']] == null) {
-                    privilege_2_iam_mapping_name[action['action']] = [];
+                if (privilege_to_iam_mapping_name[action['action']] == null) {
+                    privilege_to_iam_mapping_name[action['action']] = [];
                 }
-                privilege_2_iam_mapping_name[action['action']].push(iam_mapping_name)
+                privilege_to_iam_mapping_name[action['action']].push(iam_mapping_name)
             }
         }
     }
-    let used_by_methods = (privilege_2_iam_mapping_name[privilege] || []).map(iam_mapping_name => {
+    let used_by_methods = (privilege_to_iam_mapping_name[privilege] || []).map(iam_mapping_name => {
         return "<a href=\"/api/" + sdk_map['sdk_method_iam_mappings'][iam_mapping_name][0]['action'].split(":")[0] + "#" + iam_mapping_name.replace(".", "_") + "\">" + iam_mapping_name + "</a>";
     });