I found the `getUsedBy` function is compute-heavy since it iterates `sdk_map['sdk_method_iam_mappings']` many times.
By caching the `action`-to-`iam_mapping_name` mapping, `/tag/*` and `/iam/*` pages are now very fast.

Page | Before | After
---|--:|--:
https://aws.permissions.cloud/tag/aws:RequestTag/$%7BTagKey%7D | 10,768 ms | 11 ms
https://aws.permissions.cloud/iam/dynamodb | 487 ms | 10 ms
https://aws.permissions.cloud/iam/s3 | 860 ms | 10 ms